Cache the result of Markdown text
