+++ Rocket.Chat/Views/Cells/Chat/ChatMessageCell.swift
@@ -185,21 +185,13 @@ final class ChatMessageCell: UICollectionViewCell {
             labelUsername.text = message.user?.username
         }
 
-        let text = NSMutableAttributedString(string: message.textNormalized())
+        let text = MessageTextCacheManager.shared.message(for: message)
 
-        if self.message.isSystemMessage() {
-            text.setFont(MessageTextFontAttributes.italicFont)
-            text.setFontColor(MessageTextFontAttributes.systemFontColor)
-        } else {
-            text.setFont(MessageTextFontAttributes.defaultFont)
-            text.setFontColor(MessageTextFontAttributes.defaultFontColor)
-        }
-
-        if self.message.temporary {
+        if message.temporary {
             text.setFontColor(MessageTextFontAttributes.systemFontColor)
         }
 
-        labelText.attributedText = text.transformMarkdown()
+        labelText.attributedText = text
 
         insertGesturesIfNeeded()
         insertAttachments()
